# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8088
  description: Generated server url
paths:
  /mytask/createUser:
    post:
      tags:
      - task-controller
      summary: Creates a new User
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDto"
        required: true
      responses:
        "201":
          description: Created User
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
  /mytask/retriveUser/{id}:
    get:
      tags:
      - task-controller
      summary: Retrive User
      operationId: getByUserId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Ok User retrived
          content:
            '*/*':
              schema:
                type: object
  /mytask/healthCheck/{value}:
    get:
      tags:
      - task-controller
      operationId: healthCheck
      parameters:
      - name: value
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /mytask/getAllUsers:
    get:
      tags:
      - task-controller
      summary: Get All  Users
      operationId: AllUsers
      responses:
        "200":
          description: Ok Retrived All Users
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
components:
  schemas:
    UserDto:
      required:
      - email
      - name
      type: object
      properties:
        userId:
          type: integer
          format: int32
        name:
          maxLength: 32
          minLength: 8
          type: string
        email:
          type: string
        password:
          maxLength: 12
          minLength: 8
          type: string
    User:
      required:
      - email
      - name
      type: object
      properties:
        userId:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
        password:
          maxLength: 12
          minLength: 8
          type: string